Port NaN fixes form libcircllhist
It turns out that these changes do not affect the functionality. The bugs
observed in libcircllhist were caused by `(int_8) 0xff == -1` which is
within the -99 .. 99 range. In JS we have `0xff == 255` which lands outside
and is thus correctly classified as NaN.

We choose to port the changes over regardless to keep the code consistent.